+++ nrf_802154/serialization/include/host/nrf_802154.h
@@ -831,6 +831,13 @@ uint64_t nrf_802154_time_get(void);
 void nrf_802154_security_global_frame_counter_set(uint32_t frame_counter);
 
 /**
+ * @brief Sets nRF 802.15.4 Radio Driver MAC Global Frame Counter if the value passed is larger than current.
+ *
+ * @param[in] frame_counter Frame counter to set.
+ */
+void nrf_802154_security_global_frame_counter_set_if_larger(uint32_t frame_counter);
+
+/**
  * @brief Store the 802.15.4 MAC Security Key inside the nRF 802.15.4 Radio Driver.
  *
  * @param[in] p_key Pointer to the key to store. Refer to @ref nrf_802154_key_t for details.