+++ nrf_802154/sl/include/nrf_802154_sl_periphs.h
@@ -41,11 +41,6 @@
 #define NRF_802154_SL_PERIPHS_H__
 
 #include <nrfx.h>
-#if defined(DPPI_PRESENT)
-#include "hal/nrf_dppi.h"
-#else
-#include "hal/nrf_ppi.h"
-#endif
 
 #if NRF_802154_USE_INTERNAL_INCLUDES
 #include "nrf_802154_sl_periphs_internal.h"